[Data] Improved loading from column of URIs (#55824)
## Why are these changes needed?
There are many use cases for Ray Data where you might have a column of
URIs, that you want to transform into a column of bytes corresponding to
the bytes in those URIs. Currently Ray Data does, not handle this use
case very well because the incoming number of bytes is quite small
relative to the size of the bytes loaded when loading the data from the
URIs. This introduces a new code path that can be used to estimate the
size of the files, allowing for more adequate block sizes to be used
leading to increased performance. In followups, we will introduce
additional code to allow users to efficiently decode the bytes that have
been loaded from the URIs.

<!-- Please give a short summary of the change and the problem this
solves. -->

## Related issue number

<!-- For example: "Closes #1234" -->

## Checks

- [ ] I've signed off every commit(by using the -s flag, i.e., `git
commit -s`) in this PR.
- [ ] I've run `scripts/format.sh` to lint the changes in this PR.
- [ ] I've included any doc changes needed for
https://docs.ray.io/en/master/.
- [ ] I've added any new APIs to the API Reference. For example, if I
added a
method in Tune, I've added it in `doc/source/tune/api/` under the
           corresponding `.rst` file.
- [ ] I've made sure the tests are passing. Note that there might be a
few flaky tests, see the recent failures at https://flakey-tests.ray.io/
- Testing Strategy
   - [ ] Unit tests
   - [ ] Release tests
   - [ ] This PR is not tested :(

---------

Signed-off-by: Matthew Owen <mowen@anyscale.com>